Adding namespaces.json and schema to Item page views, Organization of front-end files, External References ongoing
