+++ base.ini
@@ -14,6 +14,7 @@ aws_ip_ranges_path = %(here)s/aws-ip-ranges.json
 #this is to reroute downloads to something other than aws s3
 #download_proxy = https://download.encodeproject.org/
 annotations_path = %(here)s/annotations.json
+namespaces_path = %(here)s/src/encoded/schemas/namespaces.json
 
 # Only run ec2metadata on ec2 instances
 # XXX really need to reorganise ini files for more reuse
