+++ src/encoded/schemas/namespaces.json
@@ -1,5 +1,17 @@
 {
-    "UniProtKB": "http://www.uniprot.org/uniprot/",
-    "SRA": "https://www.ncbi.nlm.nih.gov/sra/?term=",
-    "GEO": "http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc="
+    "title": "External Reference Namespaces",
+    "id": "/profiles/namespaces.json",
+    "$schema": "http://json-schema.org/draft-04/schema#",
+    "additionalProperties": false,
+    "mixinProperties": [
+        { "$ref": "mixins.json#/schema_version" },
+        { "$ref": "mixins.json#/uuid" },
+        { "$ref": "mixins.json#/submitted" }
+    ],
+    "uris": {
+        "UniProtKB": "http://www.uniprot.org/uniprot/",
+        "SRA": "https://www.ncbi.nlm.nih.gov/sra/?term=",
+        "GEO": "http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc="
+    }
+    
 }
