+++ src/encoded/static/components/item-pages/components/index.js
@@ -1,0 +1,11 @@
+'use strict';
+
+/**
+ * A directory of methods and maybe a mini-component or two for common use.
+ */
+
+
+module.exports.FormattedInfoBlock = require('./FormattedInfoBlock');
+module.exports.ItemHeader = require('./ItemHeader');
+module.exports.PartialList = require('./PartialList');
+module.exports.ExternalReferenceLink = require('./ExternalReferenceLink');
